test(gemini): stabilize Stage 5 PostgreSQL matrices
Fix host-only Stage 5 matrix failures: explicitly type the requeue timestamp as bigint, reset the synthetic leader between independent boundary cases, avoid mutating already-terminal apply rows, and measure fairness over bounded progress rather than requiring strict adjacent alternation. No runtime provider request or paid call is involved.\n\nChecks: isolated lifecycle GREEN; targeted matrices rerun during root-cause iteration.\n\nChecklist: no new public/alert contract.

+++ crates/registry/src/pg/gemini_batch_claims.rs
@@ -479,7 +479,7 @@ impl PgStore {
         Self::assert_owner_locked(&mut tx, owner, ts)?;
         let changed = tx.execute(
             "UPDATE gemini_batch_items
-                SET state='queued',next_attempt_ts=GREATEST($7,0),worker_instance=NULL,
+                SET state='queued',next_attempt_ts=GREATEST($7::bigint,0::bigint),worker_instance=NULL,
                     worker_epoch=NULL,lease_until=NULL,selected_profile_id=NULL,updated_ts=$8
               WHERE job_id=$1 AND item_index=$2 AND request_id=$3
                 AND worker_instance=$4 AND worker_epoch=$5 AND claim_generation=$6
